refactor(ui): rename packages/shared to packages/ui
Renamed the directory and its npm identity from @repo/shared to @repo/ui, along
with the Vite library build name and the package's own README. This package
has only ever contained a component library - no non-UI utilities live here -
so the name should say what it actually is, and match the common monorepo
convention of a dedicated ui package for a shared design system.
