+++ .github/.release-please-manifest.json
@@ -1,4 +1,4 @@
 {
   "apps/web": "0.0.0",
-  "packages/shared": "0.0.0"
+  "packages/ui": "0.0.0"
 }
