+++ .github/release-please-config.json
@@ -9,10 +9,10 @@
       "tag-separator": "-",
       "component": "web"
     },
-    "packages/shared": {
+    "packages/ui": {
       "include-component-in-tag": true,
       "tag-separator": "-",
-      "component": "shared"
+      "component": "ui"
     }
   }
 }
