+++ .github/workflows/rollback.yml
@@ -9,7 +9,7 @@ on:
         type: choice
         options:
           - web
-          - shared-storybook
+          - ui-storybook
       version:
         description: Version to roll back to (e.g. "1.2.3")
         required: true
@@ -51,7 +51,7 @@ jobs:
 
           # TODO: add your rollback logic here.
           # At this point the following variables are available:
-          #   APP     — the app or package-storybook name (e.g. "web" or "shared-storybook")
+          #   APP     — the app or package-storybook name (e.g. "web" or "ui-storybook")
           #   IMAGE   — the GHCR image name
           #   VERSION — the version to roll back to (e.g. "1.2.3")
           echo "❌ Rollback logic not yet implemented"
