+++ packages/ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/shared",
+  "name": "@repo/ui",
   "private": true,
   "version": "0.0.0",
   "type": "module",
